Fixed the problem with @PrimaryKey by making the POJO and the method references to the POJO, non-static. Fixed the problem with annotation processor by adding annotationProcessorOptions in build.gradle.

+++ app/src/main/java/com/portablecollections/portablecollections/CollectableDao.java
@@ -51,7 +51,7 @@ public interface CollectableDao {
      * @return a {@link Cursor} of the collectable in the database.
      */
     @Query("SELECT name, description, country, city FROM collectables WHERE id = :id")
-    Cursor selectById(int id);
+    Cursor selectById(long id);
 
     /**
      * Deletes a single collectable from the database.
@@ -60,7 +60,7 @@ public interface CollectableDao {
      * @return the number of collectables deleted. This should always be {@code 1}.
      */
     @Query("DELETE FROM collectables WHERE id = :id")
-    int deleteById(int id);
+    int deleteById(long id);
 
     /**
      * Update the collectable identified by the row ID.
